requirements.txt
selenium>=4.15.0
pandas>=2.2.0
webdriver-manager>=4.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
